Outputs a synthetic survey using a simple model

make_synthetic_data() outputs a synthetic survey, generated using a simple, stochastic model of polarisation.

Usage

make_synthetic_data(
  nrow,
  ncol,
  minority = 0.5,
  polarisation = 0,
  correlation = 0.85,
  scale = 10
)

Arguments

nrow

The number of rows in the survey

ncol

The number of columns in the survey

minority

The fraction of nodes in the smaller of the two polarised groups

polarisation

The degree of polarisation among the system's agents

correlation

Probability that group item corresponds to polarisation

scale

Range of the Likert scale

Value

A data frame corresponding to a survey.

Examples

S <- make_synthetic_data(200, 8)