Illustrates how network properties vary with the similarity threshold

make_threshold_profile() outputs properties of the agent or symbolic network as a function of similarity threshold.

Usage

make_threshold_profile(data, layer)

Arguments

data

A data frame corresponding to the attitudes held by agents with respect to a number of items

layer

A string flag specifying the type of network to be extracted,

"agent" produces the network corresponding to the agents, which we assume to be rows in data

"symbolic" produces the network corresponding to the symbols, or items, which we assume to be columns in data

Value

A data frame containing properties of the agent or symbolic network as a function of the similarity threshold. In particular, it contains three columns named

threshold, the value of the similarity threshold

ad, the average degree resulting from threshold, and

lcc, the size of the largest connected component resulting from threshold

Details

Note that this routine is expensive on large graphs. We study networks over the full range of similarity thresholds [-1, 1], and as a result, produce networks that are complete at the lower limit of that range. Note that by default we will subsample the provided survey with the C++ implementation in order to avoid memory issues. We could then allow a flag that turns off the subsampling step, at the user's peril.

Examples

make_synthetic_data(20, 5)
#>    group item_1 item_2 item_3 item_4 item_5
#> 1      0      5      7      4      3      4
#> 2      0      4      2      5      6      3
#> 3      0      4     10      9      5      7
#> 4      0     10      5      7      1      2
#> 5      1      3      4      2      5      6
#> 6      0      6      5      6      7      1
#> 7      0      4      5      5      4      3
#> 8      0      5      3      6      7      4
#> 9      0      3      6      5      4      8
#> 10     1      9      6      8      7      3
#> 11     1      4      6      1      5      6
#> 12     1      8      3      6      4      4
#> 13     1      4      7      6      1      7
#> 14     0      4      8      8      5      4
#> 15     1      6      3      1      2      4
#> 16     1      6      7      5      3      3
#> 17     1      5      7      7     10      9
#> 18     0      2      6      4      2      4
#> 19     1     10      2      5      6      4
#> 20     1      4      9      4      5      4
make_threshold_profile(S, "agent")
#> Error in eval(expr, envir, enclos): object 'S' not found